[ui] Add context menu to non-global asset graphs, direction option to lineage
